Fixes Issue 1: Add support for Pub/Sub Fixes Issue 7: Add (experimental) TLS transport
Substantial work.  It also cleans up a lot.  Much that was formerly "Public"
API is now "Private", making for much simpler documentation.  There is support
for SetConfig() and GetConfig() socket APIs, which allow configuration
directives like TLS configuration to be passed to a socket.
Also, we use this for PUB/SUB.  We've also adopted some more idiomatic
Go styles, and are using Factory methods to simplify dynamic handling
of protocols or transports that need Init to be applied.  (For example when
these keep local state.)