REP: keep header for user-space inspection
The present implementation of REP strips the header off the Recv'd message,
after storing the backtrace.

I encountered problems with this when testing the following scenario:
let REQ retry each 1sec (OptionRetryTime), and let REP sleep 5..10sec
between Recv() and Reply() (SendMsg).

The REP will send the reply to the first (original) message when coded
in a loop (as in the reqrep example). It will also reply to all
duplicated (resent) messages.

Without having the header available for inspection, the REP will have to
add extra (application-level) headers to be able to detect
duplicate/resent messages.

I checked the RFC (sp-request-reply-01), and it does not specify that
the REP should strip the header off (the REQ does not).

Using the header, the REP can store received IDs in a map, using e.g.
the first and last 4 bytes of the header; and use this for duplicate
detection.