req: skip queueing on failed send
The req.resend queue is unbuffered and hence queueing into it will
block until there is a matching receiver. This is no problem for the
resender, since it is a separate goroutine.
For the sender, this can deadlock; since it is both a consumer (rq)
and producer of the req.resend queue. This would mean it is not able
to reach the 'break' statement.
This is not a critical issue, since on adding a new endpoint, a new
sender() would start; it avoids having leftover (hanging) goroutines
in a long-running application.