Fix missing defer in the scalability test.
This caused us to be able to reach bogus high values for the client
thread count.  Once fixed, we can still perform quite nicely, and we
have the test set for 10K clients, representing 20K pipe endpoints.
This proves that we handle C10K easily.  (We can push to 20K clients,
but the test starts to degrade dramatically beyond that.)