backout changes for #286, needs work
After much testing and analysis, the changes made to improve "fast recovery"
times when a REP peer disconnects while working on a task have proven to be
overly problematic.

We believe that it would still be desirable to have REQ sockets immediately
restart/replay messages when a REP peer has vanished, but the current core
design of mangos and this protocol are not conducive to doing so without
introducing either message duplication or races.  As this change was made to
improve a failure scenario that should be rare, we are taking the decision
to defer that improvement until such time as it can be made without creating
undue complexity or risk.

Note that the core semantic behavior of REQ/REP remains correct -- messages
from REQ will be retried when a REP partner disconnects *after the retry
time out expires*.  The self healing aspect of REQ/REP is not changed, simply
that it takes a little longer than we think it *could*.

A new issue (RFE) will be filed to address this in a future redesign.