fixes #2 Add support for Named Pipes on Windows
This adds support for Windows Named Pipes as ipc:// URLs that are
compatible with nanomsg and NNG.

It also includes (untested) support for SecurityDescriptor and
InputBufferSize and OutputBufferSize tunables.   It is built on
the github.com/Microsoft/go-winio library.

Note that legacy libnanomsg (all versions 1.1.3 and earlier) is
very fragile in it's handling of IPC, and assumes that senders
will only send a single atomic write.  This assumption requires
us to make an extra data copy.  (Note that NNG has no such assumptions,
and we could easily dispense with the data copy for NNG.)