Introduce different constructors for RAW sockets.
This is a BREAKING change, as the NewProtocol functions for
protocols are now removed.

The OptionRaw option still works, for now, but it's on the chopping
block.

This work is a stepping stone towards introducing contexts for
some protocols.  Its likely that for certain protocols the raw
mode may be a completely different underlying implementation in
the future.