OptionRaw becomes read-only.
NOTE: This is a breaking change!!

Now that we have dedicated constructors for raw mode sockets,
we can eliminate the ability to change the mode.  This is more
simplification for future context work.

The Device() code now checks for raw mode, rather than changing it.
If a non-raw mode socket is pased, it returns ErrNotRaw.

The tests have been updated to reflect these changes.