Substantial refactoring, and now context support for REQ protocol.
REP is probably more interesting for most folks, but REQ is a bit
more challenging, so we did that first to make sure we didn't miss
any critical points.

This has a bunch of other API breakages.

Pipe GetProp is removed and replaced with GetOption.

Some options like the queue depths that don't make sense for sockets
without queuing (i.e. REQ and in the future REP, maybe also RESPONDENT,
and half of SURVEYOR) are removed from those sockets.

Backpressure is much more immediate as there is no longer *any* buffering.
The use of condvars and mutexes instead of channels and buffering probably
means lower latency and better performance, but for some pure throughput
cases the old buffers may have been better.  The blocking behavior when
no sender is immediately available is likely to confuse some folks.

The way to got socket protocol descriptions is changed -- see the new
socket.Info() and ProtocolInfo structure.  (This was necessary as the
new ProtocolBase field is utterly different, so we don't want sockets
exposing the Protocl directly.)

The underlying APIs for Transports have changed dramatically.

So has the underlying Protocol API, but we still have most of the old
ones (only REQ is still around.)
There is code reorganization in progress too.  The