fixes #1 Dial should be synchronous by default
This makes Dial synchronous, and adds a new option, OptionDialAsynch,
which can be set on a socket or a Dialer to restore the mangos v1
behavior.

We also have fixed bugs in the REQ/REP code, and made it correctly
self-heal "quickly" from a broken socket.  (Basically if our response
pipe disconnects, we can immediately try the request somewhere else.)