Disable AppVeyor badge.
The AppVeyor infra is too unreliable to rely on for test assessments.

We still want the feedback for PRs, but having the badge gives people the wrong
idea about this project; the problem is not that mangos is unstable on Windows,
but rather that the quality of the free AppVeyor services is too poor to execute
our tests reliably.  So bad on AppVeyor, not on mangos.